LayoutTests/transforms/3d/hit-testing/hover-rotated-with-children-negative-z-expected.txt

transformed
child 1
non transformed child
child 2
Test passes if the hover state of a transformed div is activated even if there's a non transformed element behind it.

Element at 70, 70 has id "transformed": PASS
Element at 630, 130 has id "transformed": PASS
Element at 40, 130 has id "transformed": PASS
Element at 620, 270 has id "transformed": PASS
Element at 130, 100 has id "child1": PASS
Element at 200, 110 has id "child2": PASS
